change Rating to float64

+++ readingList/internal/data/books.go
@@ -11,6 +11,6 @@ type Book struct {
 	Published int       `json:"published,omitempty"`
 	Pages     int       `json:"pages,omitempty"`
 	Genres    []string  `json:"genres,omitempty"`
-	Rating    float32   `json:"rating,omitempty"`
+	Rating    float64   `json:"rating,omitempty"`
 	Version   int32     `json:"-"`
 }
